backend/requirements.txt
python-dotenv==1.0.1
Django==5.1.1
djangorestframework==3.15.2
django-environ==0.11.2
django-structlog==8.1.0
whitenoise==6.7.0
authorizenet==1.1.5
